+++ joystick1/lewansoul.h
@@ -12,10 +12,10 @@ class LewanSoul
     void setup();
 
     // ID of desired servo to move, and destination from -100 to 100
-    void moveTo(int id, int destination);
+    void moveTo(int id, float destination);
 
     // ID of desired servo to spin, and velocity from -100 to 100
-    void spinAt(int id, int velocity);
+    void spinAt(int id, float velocity);
 };
 
 #endif // lewansoul_h
